+++ conf/buster-coverage.js
@@ -11,7 +11,7 @@ module.exports.unit = {
     require('buster-istanbul')
   ],
   'buster-istanbul': {
-    outputDirectory: '.bob/report/coverage/buster-istanbul/',
+    outputDirectory: '.bob/coverage/buster-istanbul/',
     formats: ['text', 'lcov']
   }
 };